requirements.txt
# requirements.txt
Flask==3.0.3
Flask-CORS==4.0.0
mysql-connector-python==8.2.0
bcrypt==4.1.2
python-dotenv==1.0.1
gunicorn==21.2.0  # Solo si vas a desplegar en producción